requirements.txt
asgiref==3.5.0
Django==4.0.3
Pillow==9.0.1
python-decouple==3.6
python-magic==0.4.25
sqlparse==0.4.2
